CAN_CMD_MOTOR_CONTROL(3.14f,0.0f,50.0f,5.0f,0.0f,Control_ID1); 串口通信解决：发现是串口线没有连接好的问题；注意这个函数的五个参数的顺序：注意float CAN_GetDeep_Motor(int8_t Which_x) 的返回值为float类型；还有一个未解决的问题，就是串口接收的方式发送目标位置的话非常不稳定基本上无法实现

+++ BSP/drv_can.h
@@ -92,12 +92,12 @@ typedef struct {
 #define Control_ID5    5 + Control_Motor_ID
 #define Control_ID6    6 + Control_Motor_ID
 
-#define Disable_ID1_Receive 1 + Disable_Motor_ID + 0x10
-#define Disable_ID2_Receive 2 + Disable_Motor_ID + 0x10
-#define Disable_ID3_Receive 3 + Disable_Motor_ID + 0x10
-#define Disable_ID4_Receive 4 + Disable_Motor_ID + 0x10
-#define Disable_ID5_Receive 5 + Disable_Motor_ID + 0x10
-#define Disable_ID6_Receive 6 + Disable_Motor_ID + 0x10
+#define Disable_ID1_Receive  1 + Disable_Motor_ID + 0x10
+#define Disable_ID2_Receive  2 + Disable_Motor_ID + 0x10
+#define Disable_ID3_Receive  3 + Disable_Motor_ID + 0x10
+#define Disable_ID4_Receive  4 + Disable_Motor_ID + 0x10
+#define Disable_ID5_Receive  5 + Disable_Motor_ID + 0x10
+#define Disable_ID6_Receive  6 + Disable_Motor_ID + 0x10
 
 #define Able_ID1_Receive    1 + Able_Motor_ID + 0x10
 #define Able_ID2_Receive    2 + Able_Motor_ID + 0x10
@@ -127,7 +127,7 @@ void CAN_SendCurrent(int16_t current1,int16_t current2,int16_t current3,int16_t 
 void CAN_CMD_MOTOR_ENABLE(uint32_t stdid);
 void CAN_CMD_MOTOR_DISABLE(uint32_t stdid);
 void CAN_SEND_DATA(uint16_t id, int16_t current);
-uint32_t CAN_GetDeep_Motor(int8_t Which_x);
+float CAN_GetDeep_Motor(int8_t Which_x);
 void CAN_CMD_MOTOR_CONTROL(float TargetAngle,float TargetSpeed,
                            float Kp,float Kd,float TargetTorque,float stdid);
